Refactor frame module into submodules
